+++ src/domains/merchant-import/index.ts
@@ -1,0 +1,7 @@
+export { MerchantImportCommitService, type CommitContext, type CommitResult } from "./MerchantImportCommitService";
+export { ImportPlanBuilder, summarizePlan, type PlanBuilderDeps, type ExistingProductForMatch } from "./ImportPlanBuilder";
+export {
+  sourceChecksum, canTransition, canCommit, IMPORT_STATUS_TRANSITIONS,
+  type ImportStatus, type ImportItemDecision, type ImportPlanItem, type ImportPlanSummary,
+  type ImportSessionSnapshot, type ImportSourceType, type ImportSourceMode, type ImportPreviewRequest,
+} from "./types";